Add incoming activity improvements
