+++ demo-app/src/main/kotlin/io/getstream/video/android/ui/DogfoodingNavHost.kt
@@ -16,7 +16,6 @@
 
 package io.getstream.video.android.ui
 
-import android.content.Intent
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
@@ -29,7 +28,6 @@ import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
 import io.getstream.video.android.CallActivity
 import io.getstream.video.android.core.notifications.NotificationHandler
-import io.getstream.video.android.model.StreamCallId
 import io.getstream.video.android.ui.common.StreamCallActivity
 import io.getstream.video.android.ui.join.CallJoinScreen
 import io.getstream.video.android.ui.join.barcode.BarcodeScanner
@@ -60,8 +58,6 @@ fun AppNavHost(
             )
         }
         composable(AppScreens.CallJoin.route) {
-            val context = LocalContext.current
-
             CallJoinScreen(
                 navigateToCallLobby = { cid ->
                     navController.navigate(AppScreens.CallLobby.routeWithArg(cid))
@@ -76,19 +72,6 @@ fun AppNavHost(
                 },
                 navigateToBarcodeScanner = {
                     navController.navigate(AppScreens.BarcodeScanning.route)
-                },
-                navigateToIncomingCallScreen = { call ->
-                    val intent = StreamCallActivity.callIntent(
-                        context = context,
-                        cid = StreamCallId.fromCallCid(call.cid),
-                        members = emptyList(),
-                        leaveWhenLastInCall = true,
-                        action = NotificationHandler.ACTION_INCOMING_CALL,
-                        clazz = CallActivity::class.java,
-                    ).apply {
-                        flags = Intent.FLAG_ACTIVITY_NEW_TASK or Intent.FLAG_ACTIVITY_CLEAR_TASK
-                    }
-                    context.startActivity(intent)
                 },
             )
         }
