+++ demo-app/src/main/kotlin/io/getstream/video/android/ui/join/CallJoinScreen.kt
@@ -114,7 +114,6 @@ fun CallJoinScreen(
     navigateUpToLogin: (autoLogIn: Boolean) -> Unit,
     navigateToDirectCallJoin: () -> Unit,
     navigateToBarcodeScanner: () -> Unit = {},
-    navigateToIncomingCallScreen: (Call) -> Unit = {},
 ) {
     LockScreenOrientation(orientation = Configuration.ORIENTATION_PORTRAIT)
     val uiState by callJoinViewModel.uiState.collectAsState(CallJoinUiState.Nothing)
@@ -127,7 +126,6 @@ fun CallJoinScreen(
     HandleCallJoinUiState(
         callJoinUiState = uiState,
         navigateToCallLobby = navigateToCallLobby,
-        navigateToIncomingCallScreen = navigateToIncomingCallScreen,
         navigateUpToLogin = { navigateUpToLogin(true) },
     )
 
@@ -184,7 +182,6 @@ fun CallJoinScreen(
 private fun HandleCallJoinUiState(
     callJoinUiState: CallJoinUiState,
     navigateToCallLobby: (callId: String) -> Unit,
-    navigateToIncomingCallScreen: (call: Call) -> Unit,
     navigateUpToLogin: () -> Unit,
 ) {
     LaunchedEffect(key1 = callJoinUiState) {
@@ -192,10 +189,6 @@ private fun HandleCallJoinUiState(
             is CallJoinUiState.JoinCompleted -> navigateToCallLobby.invoke(callJoinUiState.callId)
 
             is CallJoinUiState.GoBackToLogin -> navigateUpToLogin.invoke()
-
-            is CallJoinUiState.IncomingCall -> navigateToIncomingCallScreen.invoke(
-                callJoinUiState.call,
-            )
 
             else -> Unit
         }
